requirements.txt
asttokens==2.2.1
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7        
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4        
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0        
executing==1.2.0
fuzzywuzzy==0.18.0       
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1        
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
langcodes==3.3.0
langdetect==1.0.9
language-tool-python==2.7.1
Levenshtein==0.21.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
platformdirs==3.5.1
preshed==3.0.8
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-Levenshtein==0.21.0
pytz==2023.3
pywin32==306
pyzmq==25.0.2
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.30.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
thinc==8.1.10
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wasabi==1.1.1
wcwidth==0.2.6
zipp==3.15.0